Many bugfixes and added two new actions: Read from bucket and CreateBranch

+++ src/arch/data/types/TByte.java
@@ -10,7 +10,7 @@ public class TByte extends SimpleData {
 
 	int value;
 
-	public TByte(byte b) {
+	public TByte(int b) {
 		value = b;
 	}
